impl

+++ globox-dubbo-api/src/main/java/com/unlimited/sports/globox/dubbo/user/dto/UserInfoDto.java
@@ -4,7 +4,6 @@ import lombok.Data;
 
 import java.io.Serial;
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 /**
  * 用户信息DTO（RPC专用出参）
@@ -39,5 +38,5 @@ public class UserInfoDto implements Serializable {
     /**
      * 用户网球水平（NTRP）
      */
-    private BigDecimal userNtrpLevel;
+    private Double userNtrpLevel;
 }
