+++ globox-model/src/main/java/com/unlimited/sports/globox/model/auth/vo/UserInfoVo.java
@@ -9,7 +9,6 @@ import lombok.NoArgsConstructor;
 
 import java.io.Serial;
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 /**
  * 用户信息VO（用于RPC返回）
@@ -50,5 +49,5 @@ public class UserInfoVo implements Serializable {
     /**
      * 用户网球水平（NTRP）
      */
-    private BigDecimal userNtrpLevel;
+    private Double userNtrpLevel;
 }
